Update world.sh
Changed the rates to 5x instead of 3x. The reason for this is the leveling is supposed to be fairly quick. Having instant 80 makes it feel pointless, you didn't work to get there. This way it's very fast up until northrend quests, then it slows down but not too much.