XML
Added the ability to read configuration values from a XML-file instead of editing the bash scripts. Running the script without an existing XML-file will generate one for you that you can edit.

Changed the for loop for the includes to make it.. well.. actually work.

Fixed the issue in the packages include.