Start & stop
Added the ability to start and stop the servers. You can start them separately or together. If you try to stop the world server, it will send saveall to the screen, sleep for 3 seconds and then stop it.

It will not start the server if it's already running, and it will not try to stop it if it's not running.