Backup
Added the ability to perform a backup of specified databases.

Changed packages to install only the required packages for the backup function. This is good because if the system used to perform backups isn't the one running the server, then it'll install useless packages on that system.

I included the crontab settings I use on my dedicated backup system that performs backups once every hour.

Changed the name of the build function to better fit with its actions.