Quest loot
Fixed several issues I encountered with flags not being set on all quest items. The reason is that not all quest items are 0, so this should fix that issue. It will still not fix items that you loot in order to combine into a quest item but those will be added when found.

This function should only be called when the database is newly imported, so the +2048 shouldn't cause issues.