Initial commit
Initial commit of the Powershell version of this script. It will download, or update, the source code and compile it. It will copy the executable, configuration files and library files to a specific directory. It will also download the client data files and unpack them into the same directory. That is how far it's come though, as it will not update the configuration or anything like that. It can however be used already to at least handle the server to a certain extent. The client file download takes quite a while, not sure why but I don't really care either to be honest. However, if the files are already there it doesn't hurt that much since it only has to be done once.