Modules
Add these back in. It might be able to compile, but for now I'm personally leaving them as disabled. I found a lot of my issues were database-related and not core-related and I  have fixed these issues. If it compiles, great. If not, then we wait for a fix from the AHBot module.