CentOS 8
Added support for CentOS 8. I will not be running this script on CentOS myself, so if there are any issues that I haven't run into yet I won't know unless someone told me about it

The next step is to add support for Fedora 33, which is fairly similar to CentOS, but I'm really in no rush to do so

Removed the requirement of the Cameras folder when checking for client data files. I'm not even sure if this folder is actually required by the core, but it has no reason to be in here since if the other folders are missing it will grab them all anyway