LFG Loot Mode
Changed the ability to allow changing loot mode while in a LFG to be a configuration option

Normally, if a full group of players queue for a dungeon they are able to change the loot mode. However, this is not how it currently works with any available core that I have seen

Enabling this feature on a public server is obviously not recommended since the player that gets to be the leader of the dungeon group will be able to control the loot. For my purposes it is ideal though