Revert
Reverting support for both CentOS and Debian. The reason for this is that the script is, with a lack of better words, a hot mess. Using manual usage it appears to be fine, but using crontab there are just so many issues that I decided it's simply better to stick with what works. The main reason for the issues is CMake, and I simply don't have the time or energy to spend on fixing it. I'm going back to Ubuntu containers