Progression!
Started working on a set of SQL-files to manage progression of a server. It starts at initial release of World of Warcraft. I have begun going through it, but there are so much to do. I'm still working on setting level and health of things like city guards and bosses. There are also a ton of creatures that were added in later patches that I don't know about and hopefully I can get some help to go through all of this to get a proper system of progression going. The plan is to go all the way to 3.3.5 of course, which will take a long time

I only added the beginning of Patch 1.2 and 1.3 to get away from the health/level updates a little bit. Kazzak needs to be scripted but he's there, the level, health and mana is set along with his waypoints. I am using values from vMaNGOS as a guide

I have yet to test if the dungeon finder lets you actually join a group since all dungeons are disabled. This feature is however up to the server owner to disable until it's actually available in the progression

I'm also including a timeline showing what patch added what