Experienced
Add support for my module called Experienced. The module multiplies the experience, reputation and money rates based on the characters level. It also modifies it based on the day of the week.

Level 1-60 gets 5x rates, level 61-70 gets 2.5x and 71-80 gets 1x.

On weekends (friday, saturday, sunday) the value is doubled. This means that level 1-60 gets 10x on weekends.