Rates
Changed handling character rates dramatically.

Created a function that calculates the multiplier to avoid having to change the code in several places when I need to change it.

Added the ability to enable weekend rates. The idea is that the multiplier increases on fridays, saturdays and sundays.

Added the ability to always set a characters weapon skills to their maximum value when leveling up. It's possible to set the max level when this happens.