Convenience
Added all quested spells to their class trainers.

Because the quested spells of Shaman require their corresponding totem, Shamans can now obtain these using the assistant. It will not give a character extra totems, and it will not give a character totems if they have the Totem of the Earthen Ring item.

I will use the category miscellaneous in the future but until then the Shaman totems are the only thing in there.