Money on Kill
A player will be rewarded with a specified amount of copper when killing a creature. Supports groups and raids, based on the settings. Copper value has to be changed to fit your demands.

I'm not super pleased with it yet, it does check the distance from the player who killed the creature and the level of both the member and the killer. It should be able to check every member of the party or raid and only award money if no member is above the maximum level compared to the creature.