Assistant
Reworked the order of gossip options.
Added the base for players being able to get full green quested items for their selected specialization. This feature requires the player to be level 80 and obviously that the feature is enabled. The idea is to support instant level 80 servers. It's still early in the works.