Fixes
Only print information about the assistant if the feature is actually enabled.
Only change the spawn point of new characters if it is enabled.

Added a variable to completely disable the assistant so a user doesn't have to go through and disable every function if they want to turn it off.