Gems
Finally got around to starting the work on gems.

Added meta gems. It's not ideal since the player can't know what the gem is before selecting it, but if there is a way to show some kind of tooltip to show information about a gem I'll add it in the future.