Fixes
Fixed: A player shouldn't learn a higher riding skill without having the previous one.
Fixed: A player shouldn't learn mounts they do not have the skill for.
Added: Command '.learnspell' to learn everything they can. This is useful if, for instance, a player staying on a certain level (like 80) and change their talents and need to learn new talent ranks.