Max amount of files
Added the ability to limit the max amount of files that will be saved. I ran into some issues where a backup server with a smaller disk filled up and decided to implement this to help out.

My crontab is set to hourly, so I set max files to 72 to save a total of 3 days of backups.