Equipment
Rewrite of the equipment part of the assistant. It reduces the amount of code and makes it easier for me to add variants of equipment.

I will most likely add a boolean variable that lets the user switch between the current item level ~174 green/uncommon equipment and item level ~187 blue/rare equipment.